商家入驻审核查询商家相关的产品类目

+++ membercenter-biz/src/main/java/com/yimayhd/membercenter/manager/MerchantItemCategoryManager.java
@@ -53,7 +53,7 @@ public class MerchantItemCategoryManager {
     @Autowired
     ExamineDetailDOMapper examineDetailDOMapper;
 
-    public List<MerchantItemCategoryDO> findMerchantItemCategoryByMerchant(int domainId, int sellerId) {
+    public List<MerchantItemCategoryDO> findMerchantItemCategoryByMerchant(int domainId, long sellerId) {
         return merchantItemCategoryDao.selectMerchantItemCategoriesByMerchant(domainId, sellerId);
     }
